* 100.0, not 100.
git-svn-id: file:///home/cbm/dev/rpgtoolkit/svn/trunk@1362 8ad37bdc-5837-4682-a52a-fbb6d9f1e9f7

+++ vc/trans3/plugins/Callbacks.cpp
@@ -1471,7 +1471,7 @@ STDMETHODIMP CCallbacks::CBGetFighterChargePercent(int partyIdx, int fighterIdx,
 	LPFIGHTER p = getFighter(partyIdx, fighterIdx);
 	if (p)
 	{
-		*pRet = p->charge / p->chargeMax * 100;
+		*pRet = int(p->charge / double(p->chargeMax) * 100.0);
 	}
 	else
 	{
